requirements.txt
flask==3.0.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
faiss-cpu==1.7.4
pypdf==3.17.4
azure-core==1.29.5
gunicorn==21.2.0